Fix remaining phpcs and PHPUnit failures
- Order privacy provider interfaces alphabetically
- Replace dashed section comments in observer with sentence-style comments
- Remove MOODLE_INTERNAL guard from lib.php, db/uninstall.php and lang file (no side effects)
- Sort lang strings alphabetically and remove section comments
- Cast DB-returned attemptsleft to int for assertSame
- Call observer directly in completion test (redirectEvents suppresses observers)

+++ tests/manager_test.php
@@ -112,12 +112,12 @@ final class manager_test extends \advanced_testcase {
         set_config('max_attempts', 10, 'local_emailchangeconfirm');
         [$user, $request] = $this->make_request();
 
-        $previous = manager::get_pending_request($user->id)->attemptsleft;
+        $previous = (int) manager::get_pending_request($user->id)->attemptsleft;
         $this->assertSame(10, $previous);
 
         for ($i = 0; $i < 5; $i++) {
             manager::validate_token($user->id, 'wrongtoken' . str_repeat('0', 30));
-            $current = manager::get_pending_request($user->id)->attemptsleft;
+            $current = (int) manager::get_pending_request($user->id)->attemptsleft;
             $this->assertSame($previous - 1, $current,
                 'Each failed attempt must decrement attemptsleft by exactly 1.');
             $previous = $current;
